Moved the display preference code directly into Lightning Media.
