+++ modules/lightning_features/lightning_media/lightning_media.services.yml
@@ -3,3 +3,9 @@ services:
     class: '\Drupal\lightning_media\MediaHelper'
     arguments:
       - '@entity_type.manager'
+
+  lightning_media.route_subscriber:
+    class: '\Drupal\lightning_media\Routing\RouteSubscriber'
+    tags:
+      -
+        name: event_subscriber
