--- modules/lightning_features/lightning_media/modules/lightning_media_image/lightning_media_image.services.yml
@@ -1,5 +1,0 @@
-services:
-  lightning_media_image.route_subscriber:
-    class: '\Drupal\lightning_media_image\Routing\RouteSubscriber'
-    tags:
-      - { name: event_subscriber }
