+++ modules/lightning_features/lightning_media/src/Routing/RouteSubscriber.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Drupal\lightning_media_image\Routing;
+namespace Drupal\lightning_media\Routing;
 
 use Drupal\Core\Routing\RouteSubscriberBase;
-use Drupal\lightning_media_image\Form\EntityEmbedDialog;
+use Drupal\lightning_media\Form\EntityEmbedDialog;
 use Symfony\Component\Routing\RouteCollection;
 
 class RouteSubscriber extends RouteSubscriberBase {
@@ -12,10 +12,9 @@ class RouteSubscriber extends RouteSubscriberBase {
    * {@inheritdoc}
    */
   protected function alterRoutes(RouteCollection $collection) {
-    $route = $collection->get('entity_embed.dialog');
-    if ($route) {
-      $route->setDefault('_form', EntityEmbedDialog::class);
-    }
+    $collection
+      ->get('entity_embed.dialog')
+      ->setDefault('_form', EntityEmbedDialog::class);
   }
 
 }
